Moved views up to the folder structure
