+++ src/components/header/Header.css
@@ -58,7 +58,6 @@
 	@apply text-base
 			font-normal
 			text-center
-			italic
 			leading-relaxed;
 }
